Ransomware_PoC/requirements.txt
pycryptodome 
os
requests 
